requirements.txt
aiohttp
async-timeout
discord.py
pydub
pymongo
